Run solc prettier

+++ contracts/DecentralizedAutonomousTrust.sol
@@ -12,52 +12,40 @@ import "./ContinuousOffering.sol";
  * solution using this contract, you might be interested in Fairmint's
  * offering. Do not hesitate to get in touch with us: https://fairmint.co
  */
-contract DecentralizedAutonomousTrust is ContinuousOffering
-{
-  event Close(
-    uint _exitFee
-  );
-  event Pay(
-    address indexed _from,
-    uint _currencyValue
-  );
+contract DecentralizedAutonomousTrust is ContinuousOffering {
+  event Close(uint256 _exitFee);
+  event Pay(address indexed _from, uint256 _currencyValue);
   event UpdateConfig(
     address _whitelistAddress,
     address indexed _beneficiary,
     address indexed _control,
     address indexed _feeCollector,
-    uint _revenueCommitmentBasisPoints,
-    uint _feeBasisPoints,
-    uint _minInvestment,
-    uint _minDuration
+    uint256 _revenueCommitmentBasisPoints,
+    uint256 _feeBasisPoints,
+    uint256 _minInvestment,
+    uint256 _minDuration
   );
 
   /// @notice The revenue commitment of the organization. Defines the percentage of the value paid through the contract
   /// that is automatically funneled and held into the buyback_reserve expressed in basis points.
   /// Internal since this is n/a to all derivative contracts.
-  function revenueCommitmentBasisPoints() public view returns (uint)
-  {
+  function revenueCommitmentBasisPoints() public view returns (uint256) {
     return __revenueCommitmentBasisPoints;
   }
 
   /// Close
 
-  function estimateExitFee(
-    uint _msgValue
-  ) public view
-    returns(uint)
-  {
-    uint exitFee;
+  function estimateExitFee(uint256 _msgValue) public view returns (uint256) {
+    uint256 exitFee;
 
-    if(state == STATE_RUN)
-    {
-      uint reserve = buybackReserve();
+    if (state == STATE_RUN) {
+      uint256 reserve = buybackReserve();
       reserve = reserve.sub(_msgValue);
 
       // Source: t*(t+b)*(n/d)-r
       // Implementation: (b n t)/d + (n t^2)/d - r
 
-      uint _totalSupply = totalSupply();
+      uint256 _totalSupply = totalSupply();
 
       // Math worst case:
       // MAX_BEFORE_SQUARE * MAX_BEFORE_SQUARE/2 * MAX_BEFORE_SQUARE
@@ -74,12 +62,9 @@ contract DecentralizedAutonomousTrust is ContinuousOffering
         buySlopeDen
       );
       // Math: this if condition avoids a potential overflow
-      if(exitFee <= reserve)
-      {
+      if (exitFee <= reserve) {
         exitFee = 0;
-      }
-      else
-      {
+      } else {
         exitFee -= reserve;
       }
     }
@@ -93,12 +78,10 @@ contract DecentralizedAutonomousTrust is ContinuousOffering
   /// what appears to be required and any remainder will be returned to your account.  This is
   /// because another user may have a transaction mined which changes the exitFee required.
   /// For other `currency` types, the beneficiary account will be billed the exact amount required.
-  function close() public payable
-  {
-    uint exitFee = 0;
+  function close() public payable {
+    uint256 exitFee = 0;
 
-    if(state == STATE_RUN)
-    {
+    if (state == STATE_RUN) {
       exitFee = estimateExitFee(msg.value);
       _collectInvestment(msg.sender, exitFee, msg.value, true);
     }
@@ -111,17 +94,14 @@ contract DecentralizedAutonomousTrust is ContinuousOffering
 
   /// @dev Pay the organization on-chain.
   /// @param _currencyValue How much currency which was paid.
-  function pay(
-    uint _currencyValue
-  ) public payable
-  {
+  function pay(uint256 _currencyValue) public payable {
     _collectInvestment(msg.sender, _currencyValue, msg.value, false);
     require(state == STATE_RUN, "INVALID_STATE");
     require(_currencyValue > 0, "MISSING_CURRENCY");
 
     // Send a portion of the funds to the beneficiary, the rest is added to the buybackReserve
     // Math: if _currencyValue is < (2^256 - 1) / 10000 this will not overflow
-    uint reserve = _currencyValue.mul(__revenueCommitmentBasisPoints);
+    uint256 reserve = _currencyValue.mul(__revenueCommitmentBasisPoints);
     reserve /= BASIS_POINTS_DEN;
 
     // Math: this will never underflow since revenueCommitmentBasisPoints is capped to BASIS_POINTS_DEN
@@ -132,8 +112,7 @@ contract DecentralizedAutonomousTrust is ContinuousOffering
 
   /// @notice Pay the organization on-chain without minting any tokens.
   /// @dev This allows you to add funds directly to the buybackReserve.
-  function () external payable
-  {
+  function() external payable {
     require(address(currency) == address(0), "ONLY_FOR_CURRENCY_ETH");
   }
 
@@ -142,16 +121,29 @@ contract DecentralizedAutonomousTrust is ContinuousOffering
     address payable _beneficiary,
     address _control,
     address payable _feeCollector,
-    uint _feeBasisPoints,
-    uint _revenueCommitmentBasisPoints,
-    uint _minInvestment,
-    uint _minDuration
-  ) public
-  {
-    _updateConfig(_whitelistAddress, _beneficiary, _control, _feeCollector, _feeBasisPoints, _minInvestment, _minDuration);
+    uint256 _feeBasisPoints,
+    uint256 _revenueCommitmentBasisPoints,
+    uint256 _minInvestment,
+    uint256 _minDuration
+  ) public {
+    _updateConfig(
+      _whitelistAddress,
+      _beneficiary,
+      _control,
+      _feeCollector,
+      _feeBasisPoints,
+      _minInvestment,
+      _minDuration
+    );
 
-    require(_revenueCommitmentBasisPoints <= BASIS_POINTS_DEN, "INVALID_COMMITMENT");
-    require(_revenueCommitmentBasisPoints >= __revenueCommitmentBasisPoints, "COMMITMENT_MAY_NOT_BE_REDUCED");
+    require(
+      _revenueCommitmentBasisPoints <= BASIS_POINTS_DEN,
+      "INVALID_COMMITMENT"
+    );
+    require(
+      _revenueCommitmentBasisPoints >= __revenueCommitmentBasisPoints,
+      "COMMITMENT_MAY_NOT_BE_REDUCED"
+    );
     __revenueCommitmentBasisPoints = _revenueCommitmentBasisPoints;
 
     emit UpdateConfig(
